Add files via upload
earlier script deleted. this script finds restaurant name and its cuisine types. the code for review extracts review text, source, date. but that works only when any 1 restaurant is opened. it doesn't work in the current script, for it to work u have to change url to one of the rest and then run